Moved classes to their own files
