+++ CoreWiki.Extensibility.Common/CoreWikiModuleEvents.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CoreWiki.Extensibility.Common
+{
+    public class CoreWikiModuleEvents
+    {
+        public Action<PreSubmitArticleEventArgs> PreSubmitArticle;
+        public Action<ArticleSubmittedEventArgs> ArticleSubmitted;
+    }
+}
