+++ CoreWiki.Extensibility.Common/ExtensibilityModulesConfig.cs
@@ -1,0 +1,8 @@
+﻿namespace CoreWiki.Extensibility.Common
+{
+    public class ExtensibilityModulesConfig
+    {
+        public string Name { get; set; }
+        public string Type { get; set; }
+    }
+}
