+++ CoreWiki.Extensibility.Common/ICoreWikiModule.cs
@@ -1,0 +1,7 @@
+﻿namespace CoreWiki.Extensibility.Common
+{
+    public interface ICoreWikiModule
+    {
+        void Initialize(CoreWikiModuleEvents moduleEvents);
+    }
+}
